Move serialization helpers to @utils/data/serialization.ts
